+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - Optimized run loop slightly with (in)validation optimization.
 - Removed duplicate code in run loop.
 - Further optimized drawing loop so that image is only redrawn if any objects need repainting.
+- Removed use of RefCell from main run loop for DrawingTexture object.
 
 ## 0.3.1